+++ tomcat/src/main/java/nextstep/jwp/controller/NotMappedErrorController.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.controller;
 
 public class NotMappedErrorController extends AbstractController {
