+++ tomcat/src/main/java/nextstep/jwp/controller/ResourceController.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.controller;
 
 import java.io.IOException;
 import java.nio.file.Files;
